(TODO) Add an ExternalMemoryDescriptor and Error type to burn-tensor...
Note: This currently has issues in that it makes burn-tensor aware
of external memory sources (in this case cudarc for cuda programs).

We don't want burn-tensor to have to include all possible external sources.

The ExternalMemoryDescriptor allows us to have a streamlined API
for from_ext_data(desc, device) for all sources. desc is a single
type.